refactor(import): relocate Trakt adapter into sources/ and drop the Zip suffix
Move the migrated Trakt importer next to the other adapters under
core/import/sources/trakt/, and rename TraktZipImportService ->
TraktImportService (and its provider) to match the sibling naming — the input
format is not part of the class name, the same way MalImportService imports
XML. Add the per-source README and update the layer README: every importer is
now on the shared layer; the .xcoll/.xcollx collection-file import stays
separate by design.
